# docs/openapi/provisioner-api.yaml
openapi: 3.0.3
info:
  title: eapaka-node-provisioner API
  version: 0.3.0
  description: |
    EAP-AKA RADIUS PoC（以下「本PoC」）の Provisioning API（以下「prov」）と、aka-only-server の管理API（以下「aka」）を
    組み合わせて操作する統合API。設計は `docs/design-overview.md` を参照。
    作法は下流の 2 つの API（prov 0.4.0、aka 0.2.0）に揃えている。

    0.3.0 の変更:
    - 加入者の停止・再開（`PUT /policies/{imsi}/status`。prov の中継）を加えた。停止の印は本PoCの認可ポリシーの状態
      （`status`: `active` / `suspended`）で、鍵の置き場所によらず本PoCの Auth Server が認証を拒否する。
    - 加入者の統合リソースに `status`（認可ポリシーの状態）を加えた。作成・変更の本文では指定できない（作成は常に `active`）。

    加入者:
    - 加入者を「IMSI＋鍵の置き場所（`keyStore`）＋認可ポリシー」として扱う（`/subscribers`）。
      - `poc`（接続方式 `00`）: 鍵を本PoCの `sub:{IMSI}` に持つ（prov の `/subscribers`）。
      - `aka`（接続方式 `01`）: 鍵を aka-only-server が持つ（aka の `/subscribers`）。本PoCの vector-gateway の
        AVクライアントID（設定 `PROVISIONER_AKA_AV_CLIENT_ID`）を、加入者の許可クライアントに含める。
      - 認可ポリシーは、置き場所によらず本PoCに持つ（prov の `/policies`）。
    - 置き場所は、IMSI の PLMN と設定 `PROVISIONER_PLMN_MAP`（本PoCの `VECTOR_GATEWAY_PLMN_MAP` と同じ形式・同じ値）で決まる。
      先頭 6 桁、次に先頭 5 桁の順で照合し、`01` に当たれば `aka`、それ以外は `poc`。
    - 複数のノードにまたがる作成・変更・削除は、操作の記録（`/operations`）を残しながら行い、途中で失敗したら
      補償（作成・変更）またはやり直し（削除）で整える。

    中継:
    - RADIUSクライアント（`/clients`）、認可ポリシー（`/policies`）、セッション（`/sessions`）は prov と同じ形で中継する。
      要求の本文は検証せずにそのまま送り、下流の応答（エラーを含む）をそのまま返す。
    - 下流の監査ログ（`/prov/audit-logs`、`/aka/audit-logs`）と、vector-gateway の AVクライアント（`/aka/av-clients/{clientId}`）は、
      下流の形のまま返す（読み取りだけ）。

    認証:
...
    - 管理クライアントは全権として扱う。利用者や権限の概念はこの API にはない。
...
servers:
...
paths:
  /status:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [status]
      summary: provisioner と下流の状態を取得する
      operationId: getStatus
      description: |
        下流 2 つの `/status` と、vector-gateway の AVクライアントを確かめて返す。
        下流や Valkey に接続できなくても 200 を返し、項目ごとの `reachable` と `error` で示す。
      responses:
        '200':
          description: 状態
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /subscribers:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [subscribers]
      summary: 加入者の一覧を取得する
      operationId: listSubscribers
      description: |
        prov の加入者、aka の加入者、prov の認可ポリシーの 3 つの一覧を、同じ `cursor` と `prefix` で取り寄せ、
        IMSI の昇順にマージして返す。Ki と OPc は含まない。
        - aka の加入者は、PLMN マップで `aka` に当たる IMSI だけを対象にする（aka-only-server には本PoCと関係のない加入者もいるため）。
        - 鍵だけ、ポリシーだけの IMSI も含める（`issues` で示す）。
        - 一覧では、`poc` の IMSI について aka 側の加入者は調べない（`KEY_IN_OTHER_STORE` は個別の取得でだけ分かる場合がある）。
        - 和集合の正確な件数は求めないため、`total` は返さない。
      parameters:
        - $ref: '#/components/parameters/ImsiPrefix'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 加入者の一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
    post:
      tags: [subscribers]
      summary: 加入者を作成する
      operationId: createSubscriber
      description: |
        次の順に行う。
        1. 入力を検証する（下流と同じ規則。ポリシーの規則を含む）。
        2. prov の加入者、aka の加入者（置き場所が `aka` なら有無、`poc` なら vector-gateway の AVクライアントID を許可しているもの）、
           prov の認可ポリシーの有無を確かめる。
           どれかがあれば 409（`SUBSCRIBER_ALREADY_EXISTS`。`conflicts` にある場所）。
        3. 操作の記録を作り、置き場所に加入者（鍵）を作成する。
        4. 認可ポリシーを PUT する。

        4 が失敗したら 3 で作った加入者を削除して戻し（補償）、最初の失敗のエラーに `operationId` と `rolledBack: true` を付けて返す。
        補償も失敗したら 500（`OPERATION_INCOMPLETE`）を返し、後で自動でやり直す（`/operations`）。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriberCreate'
            examples:
              poc:
                summary: 鍵を本PoCに置く加入者
                value:
                  imsi: '001010000000001'
                  ki: 465b5ce8b199b49faa5f0a2ee238a6bc
                  opc: cd63cb71954a9f4e48a5994e37a02baf
                  policy:
                    default: deny
                    rules:
                      - nasId: '*'
                        allowedSsids: [CORP-WIFI]
              aka:
                summary: 鍵を aka-only-server に置く加入者（PLMN マップで 00102 を 01 にした場合）
                value:
                  imsi: '001020000000001'
                  keyStore: aka
                  ki: 465b5ce8b199b49faa5f0a2ee238a6bc
                  opc: cd63cb71954a9f4e48a5994e37a02baf
                  sqnType: inc32
                  policy:
                    default: allow
                    rules: []
      responses:
        '201':
          description: 作成した加入者。Ki と OPc は含まない。
          headers:
            Location:
              description: 作成した加入者の URI（例 `/admin/v1/subscribers/001010000000001`）
              schema:
                type: string
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/SubscriberConflict'
        '422':
          $ref: '#/components/responses/IdempotencyKeyMismatch'
        '500':
          $ref: '#/components/responses/InternalErrorOrIncomplete'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
  /subscribers/{imsi}:
    parameters:
      - $ref: '#/components/parameters/Imsi'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [subscribers]
      summary: 加入者を取得する
      operationId: getSubscriber
      description: |
        prov の加入者、aka の加入者、prov の認可ポリシーを並行して取得し、まとめて返す。Ki と OPc は含まない。
        3 か所のどこにもなければ 404。食い違いは `issues` で示す。
        置き場所でない方の下流にだけ接続できない場合は、`issues` に `OTHER_STORE_UNREACHABLE` を入れて 200 を返す。
      responses:
        '200':
          description: 加入者
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/SubscriberNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
    patch:
      tags: [subscribers]
      summary: 加入者を変更する
      operationId: updateSubscriber
      description: |
        指定した項目だけを変更する（JSON Merge Patch）。次の順に行う。
        1. 変更前の認可ポリシーを操作の記録に残す（`policy` を指定したとき）。
        2. 認可ポリシーを PUT する（`policy` を指定したとき。全体を置き換える）。
        3. 置き場所の加入者を PATCH する（鍵の項目を指定したとき）。

        3 が失敗したら、2 を変更前に戻し（変更前になければ削除）、最初の失敗のエラーに `operationId` と `rolledBack: true` を付けて返す。
        戻せなければ 500（`OPERATION_INCOMPLETE`）を返し、後で自動でやり直す。

        - `sqn` を指定しなければ SQN には触れない（認証で進んだ SQN を巻き戻さない）。指定した場合は、そのまま書き換える。
        - `sqnType` / `allowPlain` は `keyStore` が `aka` の加入者だけに指定できる。許可クライアント（`allowedClientIds`）は変更しない。
        - 置き場所は変更できない。
        - 鍵の項目を指定したが置き場所に加入者がない場合は 404（`USER_NOT_FOUND`）。`policy` だけなら、鍵がなくても変更できる。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/SubscriberUpdate'
            example:
              amf: b9b9
              policy:
                default: deny
                rules:
                  - nasId: AP-OFFICE-01
                    allowedSsids: [CORP-WIFI]
                    vlanId: '100'
      responses:
        '200':
          description: 変更後の加入者。Ki と OPc は含まない。
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/SubscriberNotFound'
        '409':
          $ref: '#/components/responses/OperationInProgress'
        '422':
          $ref: '#/components/responses/IdempotencyKeyMismatch'
        '500':
          $ref: '#/components/responses/InternalErrorOrIncomplete'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
    delete:
      tags: [subscribers]
      summary: 加入者を削除する
      operationId: deleteSubscriber
      description: |
        認可ポリシー、置き場所の加入者（鍵）の順に削除する（どちらも、既にないものは削除済みとして扱う）。
        削除は戻せないため、途中で失敗したら 500（`OPERATION_INCOMPLETE`）を返し、残りを後で自動でやり直す。
        - 置き場所でない方の下流にある同じ IMSI の加入者（`KEY_IN_OTHER_STORE`）は削除しない。
        - `aka` の加入者は、他の AVクライアントを許可していても、加入者ごと削除する（他の AVクライアントとは共有しない前提）。
        - 同じ IMSI のセッションは削除しない（本PoCの Provisioning API と同じ）。
        - 3 か所のどこにもなければ 404。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: 削除した
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/SubscriberNotFound'
        '409':
          $ref: '#/components/responses/OperationInProgress'
        '422':
          $ref: '#/components/responses/IdempotencyKeyMismatch'
        '500':
          $ref: '#/components/responses/InternalErrorOrIncomplete'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
  /subscribers/{imsi}/keys:
    parameters:
      - $ref: '#/components/parameters/Imsi'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [subscribers]
      summary: 加入者の Ki と OPc を取得する
      operationId: getSubscriberKeys
      description: |
        置き場所の `/subscribers/{imsi}/keys` から取得する。provisioner と置き場所の下流の監査ログに記録する（値は記録しない）。
      responses:
        '200':
          description: Ki と OPc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberKeys'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/SubscriberNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
  /operations:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [operations]
      summary: 未解決の操作の一覧を取得する
      operationId: listOperations
      description: |
        状態が `running` / `retrying` / `failed` の操作を、作成の古い順に返す。
        完了した操作（`completed` / `rolled_back` / `dismissed`）は、保持期間（既定 7 日）のあいだ `/operations/{operationId}` で取得できる。
      parameters:
        - name: status
          in: query
          description: 状態で絞り込む。
          schema:
            type: string
            enum:
              - running
              - retrying
              - failed
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: 未解決の操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /operations/{operationId}:
    parameters:
      - $ref: '#/components/parameters/OperationId'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [operations]
      summary: 操作を取得する
      operationId: getOperation
      responses:
        '200':
          description: 操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/OperationNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /operations/{operationId}/retry:
    parameters:
      - $ref: '#/components/parameters/OperationId'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    post:
      tags: [operations]
      summary: 失敗した操作をやり直す
      operationId: retryOperation
      description: |
        状態が `failed` の操作の続き（作成・変更は補償、削除は残りの削除）を、その場で 1 回行う。
        成功すれば `rolled_back`（作成・変更）または `completed`（削除）になり、失敗すれば `retrying` に戻って自動のやり直しを再開する
        （自動のやり直しをやめるまでの 24 時間を、今から数え直す）。
        本文は送らない。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: やり直した後の操作
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/OperationNotFound'
        '409':
          $ref: '#/components/responses/OperationStateConflict'
        '422':
          $ref: '#/components/responses/IdempotencyKeyMismatch'
        '500':
          $ref: '#/components/responses/InternalError'
  /operations/{operationId}/dismiss:
    parameters:
      - $ref: '#/components/parameters/OperationId'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    post:
      tags: [operations]
      summary: 失敗した操作を閉じる
      operationId: dismissOperation
      description: |
        状態が `retrying` / `failed` の操作を、手で直した後に `dismissed` にして閉じる（自動のやり直しを止める）。
        下流には何もしない。本文は送らない。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 閉じた操作
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/OperationNotFound'
        '409':
          $ref: '#/components/responses/OperationStateConflict'
        '422':
          $ref: '#/components/responses/IdempotencyKeyMismatch'
        '500':
          $ref: '#/components/responses/InternalError'
  /clients:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [clients]
      summary: RADIUSクライアントの一覧を取得する（prov の中継）
      operationId: listClients
      description: prov の `GET /clients` を中継する。共有シークレットは含まない。
      parameters:
        - name: ip
          in: query
          description: IP アドレスで絞り込む。
          schema:
            $ref: '#/components/schemas/Ipv4'
      responses:
        '200':
          description: RADIUSクライアントの一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientList'
        default:
          $ref: '#/components/responses/Relayed'
    post:
      tags: [clients]
      summary: RADIUSクライアントを登録する（prov の中継）
      operationId: createClient
      description: prov の `POST /clients` を中継する。`Location` は provisioner のパス（`/admin/v1/clients/{clientId}`）で返す。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '201':
          description: 登録した RADIUSクライアント。共有シークレットは含まない。
          headers:
            Location:
              description: 登録した RADIUSクライアントの URI（例 `/admin/v1/clients/1`）
              schema:
                type: string
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        default:
          $ref: '#/components/responses/Relayed'
  /clients/{clientId}:
    parameters:
      - $ref: '#/components/parameters/ClientId'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [clients]
      summary: RADIUSクライアントを取得する（prov の中継）
      operationId: getClient
      responses:
        '200':
          description: RADIUSクライアント
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        default:
          $ref: '#/components/responses/Relayed'
    patch:
      tags: [clients]
      summary: RADIUSクライアントを変更する（prov の中継）
      operationId: updateClient
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/ClientUpdate'
      responses:
        '200':
          description: 変更後の RADIUSクライアント
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        default:
          $ref: '#/components/responses/Relayed'
    delete:
      tags: [clients]
      summary: RADIUSクライアントを削除する（prov の中継）
      operationId: deleteClient
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: 削除した
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
        default:
          $ref: '#/components/responses/Relayed'
  /clients/{clientId}/secret:
    parameters:
      - $ref: '#/components/parameters/ClientId'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [clients]
      summary: RADIUSクライアントの共有シークレットを取得する（prov の中継）
      operationId: getClientSecret
      description: provisioner と prov の監査ログに記録する（値は記録しない）。
      responses:
        '200':
          description: 共有シークレット
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSecret'
        default:
          $ref: '#/components/responses/Relayed'
  /policies:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [policies]
      summary: 認可ポリシーの一覧を取得する（prov の中継）
      operationId: listPolicies
      parameters:
        - $ref: '#/components/parameters/ImsiPrefix'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 認可ポリシーの一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyList'
        default:
          $ref: '#/components/responses/Relayed'
  /policies/{imsi}:
    parameters:
      - $ref: '#/components/parameters/Imsi'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [policies]
      summary: 認可ポリシーを取得する（prov の中継）
      operationId: getPolicy
      responses:
        '200':
          description: 認可ポリシー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        default:
          $ref: '#/components/responses/Relayed'
    put:
      tags: [policies]
      summary: 認可ポリシーを作成する、または置き換える（prov の中継）
      operationId: putPolicy
      description: |
        prov の `PUT /policies/{imsi}` を中継する。加入者の操作と同じ IMSI のロックを取る（処理中なら 409 `OPERATION_IN_PROGRESS`）。
        同じ IMSI に未完了の操作の記録があれば 409（`OPERATION_UNRESOLVED`）。
        加入者（鍵）の有無は確認しない。`Location`（201）は provisioner のパスで返す。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyPut'
      responses:
        '200':
          description: 置き換えた認可ポリシー
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '201':
          description: 作成した認可ポリシー
          headers:
            Location:
              description: 作成した認可ポリシーの URI（例 `/admin/v1/policies/001010000000001`）
              schema:
                type: string
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        default:
          $ref: '#/components/responses/Relayed'
    delete:
      tags: [policies]
      summary: 認可ポリシーを削除する（prov の中継）
      operationId: deletePolicy
      description: |
        加入者の操作と同じ IMSI のロックを取る（処理中なら 409 `OPERATION_IN_PROGRESS`）。
        同じ IMSI に未完了の操作の記録があれば 409（`OPERATION_UNRESOLVED`）。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: 削除した
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
        default:
          $ref: '#/components/responses/Relayed'
  /policies/{imsi}/status:
    parameters:
      - $ref: '#/components/parameters/Imsi'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    put:
      tags: [policies]
      summary: 加入者を停止する、または再開する（prov の中継。0.3.0 で追加）
      operationId: putPolicyStatus
      description: |
        prov の `PUT /policies/{imsi}/status` を中継する。`suspended` にすると、本PoCの Auth Server はその加入者の認証を
        ルールを評価せずに拒否する（鍵の置き場所によらない）。`active` に戻すと再開する。効くのは次の認証からで、
        接続中のセッションは切れない。
        - 加入者の操作と同じ IMSI のロックを取る（処理中なら 409 `OPERATION_IN_PROGRESS`）。
          同じ IMSI に未完了の操作の記録があれば 409（`OPERATION_UNRESOLVED`）。
        - 本文は検証せずに送る（`status` の検証は prov。前後の空白を除き小文字にしてから検証し、`active` / `suspended`
          以外は 400）。認可ポリシーがなければ 404（`POLICY_NOT_FOUND`。prov の応答）。
        - 今と同じ状態への変更も 200。prov はこのとき書き込まず監査ログにも残さないが、provisioner は受け付けた要求として
          監査ログに残す。
        - 監査ログの操作は、応答の状態（変更後）が `suspended` なら `policy.suspend`、`active` なら `policy.resume`。
        - 停止の印は認可ポリシーにあるので、認可ポリシーを削除すると消える（作り直すと `active`）。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyStatusPut'
            example:
              status: suspended
      responses:
        '200':
          description: 状態を変えた後の認可ポリシー
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        default:
          $ref: '#/components/responses/Relayed'
  /sessions:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [monitoring]
      summary: アクティブセッションを取得する（prov の中継）
      operationId: listSessions
      parameters:
        - name: imsi
          in: query
          description: 加入者の IMSI で絞り込む。
          schema:
            $ref: '#/components/schemas/Imsi'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: アクティブセッション
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'
        default:
          $ref: '#/components/responses/Relayed'
  /audit-logs:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [monitoring]
      summary: provisioner の監査ログを取得する
      operationId: listAuditLogs
      description: |
        provisioner の監査ログを新しい順に返す。続きを取得するには、応答の `nextBefore` を次の `before` に渡す。
        件数の上限（`PROVISIONER_AUDIT_MAX`、既定 10000）を超えた古いものは消える。標準出力のログにも同じものが出る。
      parameters:
        - $ref: '#/components/parameters/AuditBefore'
        - $ref: '#/components/parameters/AuditLimit'
      responses:
        '200':
          description: 監査ログ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /prov/audit-logs:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [downstream]
      summary: prov の監査ログを取得する（中継）
      operationId: listProvAuditLogs
      description: prov の `GET /audit-logs` を中継する。
      parameters:
        - $ref: '#/components/parameters/AuditBefore'
        - $ref: '#/components/parameters/AuditLimit'
      responses:
        '200':
          description: prov の監査ログ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvAuditLogList'
        default:
          $ref: '#/components/responses/Relayed'
...
components:
  parameters:
    Imsi:
      name: imsi
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Imsi'
...
    ImsiPrefix:
      name: prefix
      in: query
      description: IMSI の前方一致条件。
      schema:
        type: string
        pattern: '^[0-9]{1,15}$'
...
    OperatorId:
      name: X-Operator-Id
      in: header
      required: false
      description: 操作者 ID。監査ログに記録し、下流にもそのまま渡す。形式が違えば 400（`OPTIONAL_IE_INCORRECT`）。
      schema:
        type: string
        pattern: '^[A-Za-z0-9._@-]{1,64}$'
...